refactor:user->memeber로 변경
