--- src/main/java/com/req2res/actionarybe/domain/auth/service/CustomUserDetailsService.java
@@ -1,25 +1,0 @@
-package com.req2res.actionarybe.domain.auth.service;
-
-import com.req2res.actionarybe.domain.user.entity.User;
-import com.req2res.actionarybe.domain.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.*;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String loginId) throws UsernameNotFoundException {
-        User u = userRepository.findByLoginId(loginId)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        return new org.springframework.security.core.userdetails.User(
-                u.getLoginId(), u.getPassword(), List.of(new SimpleGrantedAuthority("ROLE_USER"))
-        );
-    }
-}
