+++ src/main/java/com/req2res/actionarybe/domain/member/controller/MemberController.java
@@ -1,0 +1,5 @@
+package com.req2res.actionarybe.domain.member.controller;
+
+public class UserController {
+
+}
