+++ src/main/java/com/req2res/actionarybe/domain/member/entity/Member.java
@@ -1,14 +1,10 @@
-package com.req2res.actionarybe.domain.user.entity;
+package com.req2res.actionarybe.domain.member.entity;
 
 import com.req2res.actionarybe.global.Timestamped;
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 
 @Entity
