--- src/main/java/com/req2res/actionarybe/domain/user/controller/UserController.java
@@ -1,5 +1,0 @@
-package com.req2res.actionarybe.domain.user.controller;
-
-public class UserController {
-
-}
